restore poa serializer to v1 only
