+++ modules/claims_api/app/controllers/claims_api/v1/forms/power_of_attorney_controller.rb
@@ -73,7 +73,7 @@ module ClaimsApi
           end
 
           claims_v1_logging('poa_submit', message: "poa_submit complete, poa: #{power_of_attorney&.id}")
-          render json: ClaimsApi::V1::PowerOfAttorneySerializer.new(power_of_attorney)
+          render json: ClaimsApi::PowerOfAttorneySerializer.new(power_of_attorney)
         end
 
         # PUT to upload a wet-signed 2122 form.
@@ -96,7 +96,7 @@ module ClaimsApi
           # If upload is successful, then the PoaUpater job is also called to update the code in BGS.
           ClaimsApi::PoaVBMSUploadJob.perform_async(@power_of_attorney.id, 'put')
 
-          render json: ClaimsApi::V1::PowerOfAttorneySerializer.new(@power_of_attorney)
+          render json: ClaimsApi::PowerOfAttorneySerializer.new(@power_of_attorney)
         end
 
         # GET the current status of a previous POA change request.
@@ -105,7 +105,7 @@ module ClaimsApi
         def status
           find_poa_by_id
 
-          render json: ClaimsApi::V1::PowerOfAttorneySerializer.new(@power_of_attorney)
+          render json: ClaimsApi::PowerOfAttorneySerializer.new(@power_of_attorney)
         end
 
         # GET current POA for a Veteran.
