+++ modules/claims_api/spec/serializers/power_of_attorney_serializer_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe ClaimsApi::V1::PowerOfAttorneySerializer, type: :serializer do
+describe ClaimsApi::PowerOfAttorneySerializer, type: :serializer do
   include SerializerSpecHelper
 
   subject { serialize(poa_submission, serializer_class: described_class) }
